Add script which converts a dir of SVGs to react components
This was a bit of trouble for me
(https://www.npmjs.com/package/svg-to-react) ... I banged my head
against it before reading the source and discovering that I needed the
JSX lang extension so that Babylon could parse SVG (as if they're JSX
expressions). In any case the difference here is that `transform` prop.
If there are IDs or classnames on the input SVG we should be able to use
them to decide whether or not we need to transform some props, before
they go the SVG... I have some ideas about how this might go, but it
seems at least like the lowest touch way to get some flexibilty in these
SVGs.