Add a heap of SVG files from Luke
I'm not putting these in source ... I'm not sure that with this new
method we'll have a publish-time build step any more but rather we'll
check in the transformed SVGs. I think it'll be import (at least) at
first to review what the trasform is going to the source SVGs, and also
the condition of the source files.