Add first set of generated icon components
In the not too distant future it might be nice to publish only this
director, so using the packe will look like:

```
import Calendar from '@everydayhero/ui-icons/Calendar'
```

Presently we'd get this:

```
import Calendar from '@everydayhero/ui-icons/icons/Calendar'
```

The imports being modular might be a good idea going forward, or maybe
not so much. We'll see.